fix(driver_twai): added ctu_official workaround, fixed error count info

Closes IDFGH-16314, IDFGH-16278, IDFGH-16364, IDFGH-16383, IDFCI-3106, and IDFCI-3107

See merge request espressif/esp-idf!41570

+++ components/hal/include/hal/twai_hal.h
@@ -210,7 +210,7 @@ void twai_hal_start_bus_recovery(twai_hal_context_t *hal_ctx);
  * @param hal_ctx Context of the HAL layer
  * @return TX Error Counter Value
  */
-uint32_t twai_hal_get_tec(twai_hal_context_t *hal_ctx);
+uint16_t twai_hal_get_tec(twai_hal_context_t *hal_ctx);
 
 /**
  * @brief Get the value of the RX Error Counter
@@ -218,7 +218,7 @@ uint32_t twai_hal_get_tec(twai_hal_context_t *hal_ctx);
  * @param hal_ctx Context of the HAL layer
  * @return RX Error Counter Value
  */
-uint32_t twai_hal_get_rec(twai_hal_context_t *hal_ctx);
+uint16_t twai_hal_get_rec(twai_hal_context_t *hal_ctx);
 
 /**
  * @brief Check if certain HAL state flags are set
